Join alliance packet Fix costumes when removing weapon Quest listing/removing Daily quest packets Update quest icons with alliance ones

+++ java/game/tera/gameserver/network/clientpackets/C_Toggle_Task_Info_Window.java
@@ -36,7 +36,7 @@ public class C_Toggle_Task_Info_Window extends ClientPacket
 
         readByte();
 
-        panelState = QuestPanelState.valueOf(readShort());
+        panelState = QuestPanelState.valueOf(readByte());
     }
 
     @Override
